+++ baby-liminal-extension/src/backend/mod.rs
@@ -1,3 +1,4 @@
+use aleph_runtime_interfaces::snark_verifier::VerifierError::*;
 use environment::Environment as EnvironmentT;
 use executor::BackendExecutor as BackendExecutorT;
 use frame_support::{pallet_prelude::DispatchError, sp_runtime::AccountId32};
@@ -7,7 +8,6 @@ use pallet_contracts::chain_extension::{
     ChainExtension, Environment as SubstrateEnvironment, Ext, InitState,
     Result as ChainExtensionResult, RetVal,
 };
-use primitives::liminal::VerifierError::*;
 use sp_std::marker::PhantomData;
 
 use crate::{
