+++ runtime-interfaces/src/lib.rs
@@ -1,0 +1,5 @@
+// #![deny(missing_docs)]
+#![cfg_attr(not(feature = "std"), no_std)]
+
+#[cfg(feature = "liminal")]
+pub mod snark_verifier;
